fix: skip tests requiring avm v11 version in CI
