+++ tests/artifacts/state-ops/data/ITxnOpsContract.clear.teal
@@ -1,7 +1,5 @@
-#pragma version 11
-#pragma typetrack false
+#pragma version 10
 
-// @algorandfoundation/algorand-typescript/base-contract.d.ts::BaseContract.clearStateProgram() -> uint64:
-main:
+tests/artifacts/state-ops/contract.algo.ts::ITxnOpsContract.clearStateProgram:
     pushint 1 // 1
     return
