+++ tests/artifacts/state-ops/data/StateAppGlobalExContract.approval.teal
@@ -1,57 +1,71 @@
-#pragma version 11
-#pragma typetrack false
+#pragma version 10
 
-// @algorandfoundation/algorand-typescript/arc4/index.d.ts::Contract.approvalProgram() -> uint64:
-main:
+tests/artifacts/state-ops/contract.algo.ts::StateAppGlobalExContract.approvalProgram:
     bytecblock "dummy_bytes" 0x001064756d6d795f617263345f6279746573
     txn ApplicationID
     bnz main_after_if_else@2
-    // tests/artifacts/state-ops/contract.algo.ts:359
+    callsub constructor
+
+main_after_if_else@2:
+    callsub __puya_arc4_router__
+    return
+
+
+// tests/artifacts/state-ops/contract.algo.ts::StateAppGlobalExContract.constructor() -> void:
+constructor:
+    // tests/artifacts/state-ops/contract.algo.ts:349
+    // export class StateAppGlobalExContract extends arc4.Contract {
+    proto 0 0
+    // tests/artifacts/state-ops/contract.algo.ts:350
     // globalUint64 = GlobalState({ key: Bytes('global_uint64'), initialValue: Uint64(2) })
     pushbytes "global_uint64"
     pushint 2 // 2
     app_global_put
-    // tests/artifacts/state-ops/contract.algo.ts:360
+    // tests/artifacts/state-ops/contract.algo.ts:351
     // globalBytes = GlobalState({ key: Bytes('global_bytes'), initialValue: Bytes('dummy_bytes') })
     pushbytes "global_bytes"
     bytec_0 // "dummy_bytes"
     app_global_put
-    // tests/artifacts/state-ops/contract.algo.ts:361
+    // tests/artifacts/state-ops/contract.algo.ts:352
     // global_uint64_explicit = GlobalState({ initialValue: Uint64(2) })
     pushbytes "global_uint64_explicit"
     pushint 2 // 2
     app_global_put
-    // tests/artifacts/state-ops/contract.algo.ts:362
+    // tests/artifacts/state-ops/contract.algo.ts:353
     // global_bytes_explicit = GlobalState({ initialValue: Bytes('dummy_bytes') })
     pushbytes "global_bytes_explicit"
     bytec_0 // "dummy_bytes"
     app_global_put
-    // tests/artifacts/state-ops/contract.algo.ts:363
+    // tests/artifacts/state-ops/contract.algo.ts:354
     // globalArc4Bytes = GlobalState({ key: Bytes('global_arc4_bytes'), initialValue: new DynamicBytes('dummy_arc4_bytes') })
     pushbytes "global_arc4_bytes"
     bytec_1 // 0x001064756d6d795f617263345f6279746573
     app_global_put
-    // tests/artifacts/state-ops/contract.algo.ts:364
+    // tests/artifacts/state-ops/contract.algo.ts:355
     // global_arc4_bytes_explicit = GlobalState({ initialValue: new DynamicBytes('dummy_arc4_bytes') })
     pushbytes "global_arc4_bytes_explicit"
     bytec_1 // 0x001064756d6d795f617263345f6279746573
     app_global_put
+    retsub
 
-main_after_if_else@2:
-    // tests/artifacts/state-ops/contract.algo.ts:358
+
+// tests/artifacts/state-ops/contract.algo.ts::StateAppGlobalExContract.__puya_arc4_router__() -> uint64:
+__puya_arc4_router__:
+    // tests/artifacts/state-ops/contract.algo.ts:349
     // export class StateAppGlobalExContract extends arc4.Contract {
+    proto 0 1
     txn NumAppArgs
-    bnz main_after_if_else@9
+    bnz __puya_arc4_router___after_if_else@6
     txn OnCompletion
-    bnz main_after_if_else@9
+    bnz __puya_arc4_router___after_if_else@6
     txn ApplicationID
     !
     assert // can only call when creating
     pushint 1 // 1
-    return
+    retsub
 
-main_after_if_else@9:
-    // tests/artifacts/state-ops/contract.algo.ts:358
+__puya_arc4_router___after_if_else@6:
+    // tests/artifacts/state-ops/contract.algo.ts:349
     // export class StateAppGlobalExContract extends arc4.Contract {
     pushint 0 // 0
-    return
+    retsub
