+++ tests/artifacts/state-ops/data/StateAppLocalExContract.approval.teal
@@ -1,79 +1,87 @@
-#pragma version 11
-#pragma typetrack false
+#pragma version 10
 
-// @algorandfoundation/algorand-typescript/arc4/index.d.ts::Contract.approvalProgram() -> uint64:
-main:
-    // tests/artifacts/state-ops/contract.algo.ts:409
-    // export class StateAppLocalExContract extends arc4.Contract {
-    txn NumAppArgs
-    bz main_bare_routing@6
-    pushbytes 0x30c6d58a // method "opt_in()void"
-    txna ApplicationArgs 0
-    match main_opt_in_route@3
-
-main_after_if_else@10:
-    // tests/artifacts/state-ops/contract.algo.ts:409
-    // export class StateAppLocalExContract extends arc4.Contract {
-    pushint 0 // 0
+tests/artifacts/state-ops/contract.algo.ts::StateAppLocalExContract.approvalProgram:
+    intcblock 1 0
+    callsub __puya_arc4_router__
     return
 
-main_opt_in_route@3:
-    // tests/artifacts/state-ops/contract.algo.ts:414
+
+// tests/artifacts/state-ops/contract.algo.ts::StateAppLocalExContract.__puya_arc4_router__() -> uint64:
+__puya_arc4_router__:
+    // tests/artifacts/state-ops/contract.algo.ts:400
+    // export class StateAppLocalExContract extends arc4.Contract {
+    proto 0 1
+    txn NumAppArgs
+    bz __puya_arc4_router___bare_routing@5
+    pushbytes 0x30c6d58a // method "opt_in()void"
+    txna ApplicationArgs 0
+    match __puya_arc4_router___opt_in_route@2
+    intc_1 // 0
+    retsub
+
+__puya_arc4_router___opt_in_route@2:
+    // tests/artifacts/state-ops/contract.algo.ts:405
     // @arc4.abimethod({ allowActions: ['OptIn'] })
     txn OnCompletion
-    pushint 1 // OptIn
+    intc_0 // OptIn
     ==
     assert // OnCompletion is not OptIn
     txn ApplicationID
     assert // can only call when not creating
     callsub opt_in
-    pushint 1 // 1
-    return
+    intc_0 // 1
+    retsub
 
-main_bare_routing@6:
-    // tests/artifacts/state-ops/contract.algo.ts:409
+__puya_arc4_router___bare_routing@5:
+    // tests/artifacts/state-ops/contract.algo.ts:400
     // export class StateAppLocalExContract extends arc4.Contract {
     txn OnCompletion
-    bnz main_after_if_else@10
+    bnz __puya_arc4_router___after_if_else@9
     txn ApplicationID
     !
     assert // can only call when creating
-    pushint 1 // 1
-    return
+    intc_0 // 1
+    retsub
+
+__puya_arc4_router___after_if_else@9:
+    // tests/artifacts/state-ops/contract.algo.ts:400
+    // export class StateAppLocalExContract extends arc4.Contract {
+    intc_1 // 0
+    retsub
 
 
 // tests/artifacts/state-ops/contract.algo.ts::StateAppLocalExContract.opt_in() -> void:
 opt_in:
-    // tests/artifacts/state-ops/contract.algo.ts:414-415
+    // tests/artifacts/state-ops/contract.algo.ts:405-406
     // @arc4.abimethod({ allowActions: ['OptIn'] })
     // opt_in(): void {
     proto 0 0
-    // tests/artifacts/state-ops/contract.algo.ts:416
+    // tests/artifacts/state-ops/contract.algo.ts:407
     // this.localBytes(Global.creatorAddress).value = Bytes('dummy_bytes_from_external_contract')
     global CreatorAddress
-    // tests/artifacts/state-ops/contract.algo.ts:411
+    // tests/artifacts/state-ops/contract.algo.ts:402
     // localBytes = LocalState<bytes>({ key: 'local_bytes' })
     pushbytess "local_bytes" "dummy_bytes_from_external_contract" // "local_bytes", "dummy_bytes_from_external_contract"
-    // tests/artifacts/state-ops/contract.algo.ts:416
+    // tests/artifacts/state-ops/contract.algo.ts:407
     // this.localBytes(Global.creatorAddress).value = Bytes('dummy_bytes_from_external_contract')
     app_local_put
-    // tests/artifacts/state-ops/contract.algo.ts:417
+    // tests/artifacts/state-ops/contract.algo.ts:408
     // this.localUint64(Global.creatorAddress).value = Uint64(99)
     global CreatorAddress
-    // tests/artifacts/state-ops/contract.algo.ts:410
+    // tests/artifacts/state-ops/contract.algo.ts:401
     // localUint64 = LocalState<uint64>({ key: 'local_uint64' })
     pushbytes "local_uint64"
-    // tests/artifacts/state-ops/contract.algo.ts:417
+    // tests/artifacts/state-ops/contract.algo.ts:408
     // this.localUint64(Global.creatorAddress).value = Uint64(99)
     pushint 99 // 99
     app_local_put
-    // tests/artifacts/state-ops/contract.algo.ts:418
+    // tests/artifacts/state-ops/contract.algo.ts:409
     // this.localArc4Bytes(Global.creatorAddress).value = new DynamicBytes('dummy_arc4_bytes')
     global CreatorAddress
-    // tests/artifacts/state-ops/contract.algo.ts:412
+    // tests/artifacts/state-ops/contract.algo.ts:403
     // localArc4Bytes = LocalState<DynamicBytes>({ key: 'local_arc4_bytes' })
     pushbytess "local_arc4_bytes" 0x001064756d6d795f617263345f6279746573 // "local_arc4_bytes", 0x001064756d6d795f617263345f6279746573
-    // tests/artifacts/state-ops/contract.algo.ts:418
+    // tests/artifacts/state-ops/contract.algo.ts:409
     // this.localArc4Bytes(Global.creatorAddress).value = new DynamicBytes('dummy_arc4_bytes')
     app_local_put
     retsub
