+++ tests/state-op-codes.spec.ts
@@ -18,7 +18,6 @@ import { AppExpectingEffects } from './artifacts/created-app-asset/contract.algo
 import {
   ItxnDemoContract,
   ITxnOpsContract,
-  StateAcctParamsGetContract,
   StateAppGlobalContract,
   StateAppGlobalExContract,
   StateAppLocalContract,
@@ -35,10 +34,13 @@ import appLocalExAppSpecJson from './artifacts/state-ops/data/StateAppLocalExCon
 import appParamsAppSpecJson from './artifacts/state-ops/data/StateAppParamsContract.arc32.json'
 import assetHoldingAppSpecJson from './artifacts/state-ops/data/StateAssetHoldingContract.arc32.json'
 import assetParamsAppSpecJson from './artifacts/state-ops/data/StateAssetParamsContract.arc32.json'
+import { StateAcctParamsGetContract } from './artifacts/state-ops/state-acct-params-get-contract.algo'
 import {
   generateTestAccount,
   generateTestAsset,
+  getAlgodMajorVersion,
   getAlgorandAppClient,
+  getAlgorandAppClientV11,
   getAlgorandAppClientWithApp,
   getAvmResult,
   getLocalNetDefaultAccount,
@@ -47,13 +49,15 @@ import {
 
 describe('State op codes', async () => {
   const ctx = new TestExecutionContext()
+  const algodVersion = await getAlgodMajorVersion()
+  const isV11Supported = algodVersion > 4
 
   afterEach(async () => {
     ctx.reset()
   })
 
-  describe('AcctParams', async () => {
-    const [appClient, dummyAccount] = await Promise.all([getAlgorandAppClient(acctParamsAppSpecJson as AppSpec), generateTestAccount()])
+  describe.skipIf(!isV11Supported)('AcctParams', async () => {
+    const [appClient, dummyAccount] = await Promise.all([getAlgorandAppClientV11(acctParamsAppSpecJson as AppSpec), generateTestAccount()])
 
     test.each([
       ['verify_acct_balance', INITIAL_BALANCE_MICRO_ALGOS + 100_000],
@@ -86,7 +90,7 @@ describe('State op codes', async () => {
       })
 
       const avmResult = await getAvmResult(
-        { appClient, sendParams: { staticFee: AlgoAmount.Algos(1000) } },
+        { appClient: appClient!, sendParams: { staticFee: AlgoAmount.Algos(1000) } },
         methodName,
         asUint8Array(dummyAccount.bytes),
       )
@@ -109,7 +113,7 @@ describe('State op codes', async () => {
         expect(op.AcctParams.acctIncentiveEligible(mockAccount)).toEqual([true, true])
       })
 
-      await appClient.algorand.send.onlineKeyRegistration({
+      await appClient!.algorand.send.onlineKeyRegistration({
         sender: encodeAddress(asUint8Array(dummyAccount.bytes)),
         voteKey: getRandomBytes(32).asUint8Array(),
         selectionKey: getRandomBytes(32).asUint8Array(),
@@ -120,7 +124,7 @@ describe('State op codes', async () => {
         staticFee: AlgoAmount.Algos(10),
       })
       const avmResult = await getAvmResult(
-        { appClient, sendParams: { staticFee: AlgoAmount.Algos(10) } },
+        { appClient: appClient!, sendParams: { staticFee: AlgoAmount.Algos(10) } },
         'verify_acct_incentive_eligible',
         asUint8Array(dummyAccount.bytes),
       )
